pyproject.toml
[project]
name = "drfproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "coverage (>=7.10.7,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
